Fix client/server span classification
The Span object has custom logic to set `Span#isRpc` and `Span#isClient` based
on the tags passed into `Span#setTag`.
Creating a Span using SpanBuilder bypasses this logic causing `Span#isRPC` to be
false. This causes ThriftSpanConverter to use lc annotations instead of sr/ss

This commit sanitizes input from SpanBuilder so that setting tags on the
SpanBuilder results in the same behavior as setting a tag on the Span