Fix RemoteReporter problems with command queue is full
When the command queue is full it’s not possible to queue control
commands - FlushCommand or CloseCommand. They were throwing exceptions.

This change replaces BLockingQueue.add with BlockingQueue.offer to
handle this situation.

I’ve modified test InMemorySender to allow it to block. This makes it
possible to easily fill the RemoteReporter’s command queue in tests.