Handle full queue better in RemoteReporter (#180)
* Fix RemoteReport race condition when queue is full

The check “commandQueue.size() == maxQueueSize” is not thread safe, this lead to:
java.lang.IllegalStateException: Queue full

It’s better to .offer() and check the returned value

* Fix RemoteReporter problems with command queue is full

When the command queue is full it’s not possible to queue control
commands - FlushCommand or CloseCommand. They were throwing exceptions.

This change replaces BLockingQueue.add with BlockingQueue.offer to
handle this situation.

I’ve modified test InMemorySender to allow it to block. This makes it
possible to easily fill the RemoteReporter’s command queue in tests.

* RemoteReporter’s queue length metric update when queue full

When the command queue is full further FlushCommands are not enqueued
and queue length metric does not update. I moved it outside the
command - now it updates whenever scheduled flush() executes.