Automated the release process
This PR automates the release process, by setting the version to -SNAPSHOT
when on a regular build, but to a proper fixed version when building for a tag.

This effectively means that releasing is just a matter of tagging the code at
a given point.

It's still needed to manually bump the version after a tag.

At this stage, closing/releasing of the Nexus repository is still manual and
will be automated if the main release procedure proves stable enough.