Update the ReporterConfig to set Sender Correctly (#370)
A small bug in the private getReporter(..) method was resulting in the sender
always being set to the default UdpSender. This is because the field
`senderConfiguration.sender` is null until the `getSender()` method is called on
it. Because the reporter builder was accessing the field directly rather than
through the getter, it was always UdpSen a null sender, thus defaulting to the
UdpSender even if the JAEGER_ENDPOINT property was set.

This commit fixes the problem, accessing the configured sender through the
appropriate getter, allowing the user configurations to propagate through to the
tracer retrieved via the `Configuration.fromEnv().getTracer()` method.

Signed-off-by: Nate Hart <nhart@tableau.com>