Introduce JAEGER_CONFIG_MANAGER_ENDPOINT property
Introduce a new property, JAEGER_CONFIG_MANAGER_ENDPOINT, that should be
set to the URL of the Jaeger Config Manager sampling endpoint. If not
set, fall back to the now-deprecated JAEGER_SAMPLER_MANAGER_HOST_PORT as
the host and port of the URL.

If JAEGER_AGENT_HOST has been set as a property, further fall back to
that value as the host of the URL (using the default port of 5778).

This brings the SamplerConfiguration in-line with the Go client.

See also:
 * https://github.com/jaegertracing/jaeger-client-java/issues/521
 * https://github.com/jaegertracing/jaeger-client-java/issues/547
 * https://github.com/jaegertracing/jaeger-client-go/issues/326
 * https://github.com/jaegertracing/jaeger-client-go/issues/282

Signed-off-by: Luke Wood <luke@lwood.me.uk>