Tag sampler with an example implementation.
Sometimes it's important to be able to sample based on data rather than
opertaion names. Since tags are already available during the sampling
they can be used as an additional value to make a sampling decision.

For example in messages processing it would be helpful to trace some
topics more than other.

This is an example implementation of how could this work on the tracer
level.

Ideally it would be nice to have this feature on the agent level as
well.

Signed-off-by: Daniil Fedotov <fedotov.danil@gmail.com>