Do not encode span context in HTTP_HEADERS format (#721)
We know that string representation of a span context is safe to use in
HTTP header. Thus, there is no need to encode it. JS & Go clients
already avoid that unnecesserary encoding. This change is backward
compatible because all clients unencode span context even if they do not
encode it. More pragmatically, a client impacted by this change would
already be incompatible with JS & Go clients.

Signed-off-by: Clément MATHIEU <clement@unportant.info>