Fix repeatedly error logging for HttpSender (#732)
Fix problem with repeatedly error logging when HttpSender is used. The
problem was that removing "failed" command from 'commandFailedBefore'
HashSet was always done when command execution ends successfully (i.e.
without throwing SenderException). But there are some situations, when
"span is only saved in a buffer" on sender level (or there is no any
span to be sent) and remote communication is not started at all.

Unfortunatelly the solution does not solve a problem if UdpSender is
used. The reason is that PortUnreachableException is not thrown
immediately, when DatagramSocket.send(DatagramPacket) method is
invoked - this exception is thrown when subsequent call to send or
receive is done. See Javadoc:
 - DatagramSocket.send(DatagramPacket)
 - DatagramSocket.connect(InetAddress, int)

Signed-off-by: Radek Kraus <6397085+lenoch7@users.noreply.github.com>